refactor(firebase): Just import the used modules
