+++ src/setup/firebase.js
@@ -1,6 +1,9 @@
 import Vue from 'vue';
 import VueFire from 'vuefire';
-import firebase from 'firebase';
+import firebase from 'firebase/app';
+import 'firebase/auth';
+import 'firebase/database';
+import 'firebase/storage';
 
 Vue.use(VueFire);
